+++ modules/elementary_functions/functions/hankel.m
@@ -1,5 +1,5 @@
 %=============================================================================
-% Copyright (c) 2016-present Allan CORNET (Nelson)
+% Copyright (c) 2019-present Allan CORNET (Nelson)
 %=============================================================================
 % This file is part of the Nelson.
 %=============================================================================
@@ -23,35 +23,23 @@
 % License along with this program. If not, see <http://www.gnu.org/licenses/>.
 % LICENCE_BLOCK_END
 %=============================================================================
-function p = primes(n)
-  narginchk(1, 1)
-  nargoutchk(0, 1)
-  
-  if ~isscalar(n)
-    error(_('Wrong size for argument #1: scalar expected.'));
+function H = hankel(c, r)
+  narginchk(1, 2);
+  nargoutchk(0, 1);
+  c = c(:);
+  nc = length(c);
+  if nargin < 2
+    r = zeros(size(c), 'like', c);
+  elseif (~isempty(c) && ~isempty(r) && c(nc) ~= r(1))
+    warning('Nelson:hankel:AntiDiagonalConflict', _('Last element of input column does not match first element of input row.'));
   end
-  if ~isfinite(n)
-    error(_('Wrong value for argument #1: finite value expected.'));
+  r = r(:);
+  nr = length(r);
+  x = [ c; r(2:nr, 1)];
+  ij = (1:nc)' + (0:(nr-1));
+  H = x(ij);
+  if isrow(ij) && ~isempty(H)
+    H = H.';
   end
-  if ~isreal(n)
-    error(_('Wrong value for argument #1: real value expected.'));
-  end
-  
-  if n < 2
-    p = zeros(1, 0, 'like', class(n));
-    return
-  end
-  N = floor(double(n));
-  p = [1:2:N];
-  q = length(p);
-  p(1) = 2;
-  for k = 3:2:sqrt(N)
-    if p((k+1)/2)
-      p([((k*k+1)/2):k:q]) = 0;
-    end
-  end
-  p = p(find(p >= 0));
-  p = p(find(p ~= 0));
-  p = cast(p, class(n));
 end
 %=============================================================================