+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Added
 
+- `hankel` function: Hankel matrix.
 - `factor` function: Prime factors.
 - `primes` function: Prime numbers less than or equal to input value.
+- `isrow` function: Determine whether input is row vector.
+- `iscolumn` function: Determine whether input is column vector.
 
 ## Fixed
 
@@ -12,6 +15,8 @@
 ## Compilation
 
 - Boost 1.78 support (default on Windows).
+
+- CMake 3.22.1 (on Windows).
 
 # 0.5.11 (2021-11-26)